+++ components/MultiDomainsUrlManager.php
@@ -13,12 +13,12 @@
 namespace vistart\Models\components;
 
 /**
- * Description of MultipleDomainsUrlManager
+ * Description of MultiDomainsUrlManager
  *
  * @since 2.0
  * @author vistart <i@vistart.name>
  */
-class MultipleDomainsUrlManager extends \yii\web\UrlManager
+class MultiDomainsUrlManager extends \yii\web\UrlManager
 {
     //put your code here
 }